category 컴포넌트 분리, 상품등록 작업중
