+++ frontend/src/renderer/pages/register/ProductRegister.jsx
@@ -1,9 +1,35 @@
-import React from 'react';
+import { useState } from 'react';
 import plusIcon from '../../../assets/plus.png';
 import minusIcon from '../../../assets/minus.png';
 import styles from "../../../styles/register/product-register.module.css";
+import Category from "../../components/login/CategoryComponent";
+import InputForm from "../../components/login/LoginComponent";
 
 const ProductRegister = () => {
+    const [input, setInput] = useState({
+        name: "",
+        price: 0,
+        category: "",
+        explain: "",
+        state: "",
+        tradeWay: "",
+        tradePlace: ""
+    });
+
+    const setInputState = (e) => {
+        setInput({
+            ...input,
+            [e.target.name]: e.target.value
+        });
+    };
+
+    const setFreeClick = () => {
+        setInput({
+            ...input,
+            price: 0
+        });
+    };
+
     return (
         <div className={styles['form-container']}>
             <form>
@@ -48,19 +74,18 @@ const ProductRegister = () => {
                 </div>
                 <div className={styles['form-group']}>
                     <label>상품명</label>
-                    <input type="text" placeholder="상품명 입력"/>
+                    <input type="text" name='name' value={input.name} onChange={setInputState} placeholder="상품명 입력"/>
                 </div>
 
                 <div className={styles['form-group']}>
                     <label>가격</label>
                     <div className={styles['price-group']}>
-                        <input type="text" placeholder="판매가격 입력"/>
+                        <input type="text" name='price' value={input.price} onChange={setInputState} placeholder="판매가격 입력"/>
                         <div className={styles['price-free']}>
-                            <input type="checkbox" className={styles['hidden']} placeholder="판매가격 입력"/>
-                            <div className={styles['svg-1']}>
+                            <div className={styles['svg-1']} onClick={setFreeClick}>
                                 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <path className={styles['path1']} d="M23.5 12C23.5 18.3513 18.3513 23.5 12 23.5C5.64873 23.5 0.5 18.3513 0.5 12C0.5 5.64873 5.64873 0.5 12 0.5C18.3513 0.5 23.5 5.64873 23.5 12Z" stroke="#ddd" fill="#fff"></path>
-                                    <path className={styles['path2']} d="M7 12.6667L10.3846 16L18 8.5" stroke="#ddd" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"></path>
+                                    <path className={styles['path1']} d="M23.5 12C23.5 18.3513 18.3513 23.5 12 23.5C5.64873 23.5 0.5 18.3513 0.5 12C0.5 5.64873 5.64873 0.5 12 0.5C18.3513 0.5 23.5 5.64873 23.5 12Z" stroke={ input.price == 0 ? "rgb(13, 204, 90)" : "#ddd" } fill="#fff"></path>
+                                    <path className={styles['path2']} d="M7 12.6667L10.3846 16L18 8.5" stroke={ input.price == 0 ? "rgb(13, 204, 90)" : "#ddd" } strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"></path>
                                 </svg>
                             </div>
                             <span>무료나눔</span>
@@ -70,19 +95,16 @@ const ProductRegister = () => {
 
                 <div className={styles['form-group']}>
                     <label>카테고리</label>
-                    <select id='category'>
-                        <option value="">카테고리</option>
-                        <option value="1">머리</option>
-                        <option value="2">어깨</option>
-                        <option value="3">무릎</option>
-                        <option value="4">발</option>
-                        <option value="5">스웩</option>
-                    </select>
+                    <Category
+                        onChange={setInputState}
+                        input={input.category}
+                        name={'category'}
+                    />
                 </div>
 
                 <div className={styles['form-group']}>
                     <label>상품 설명</label>
-                    <textarea placeholder="상품 설명 입력"></textarea>
+                    <textarea name='explain' value={input.explain} onChange={setInputState} placeholder="상품 설명 입력"></textarea>
                 </div>
 
                 <div className={styles['form-group']}>
